[skip ci] Refactor functional implementation of hvp computation
